requirements.txt
attrs==25.3.0
beartype==0.20.2
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cn2an==0.5.23
coloredlogs==15.0.1
cryptography==44.0.2
datrie==0.8.2
demjson3==3.0.6
docopt==0.6.2
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fsspec==2025.3.2
hanziconv==0.3.2
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
onnxruntime==1.21.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pipreqs==0.4.13
proces==0.1.7
protobuf==6.30.2
pyclipper==1.3.0.post6
pycparser==2.22
pypdf==5.4.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
roman-numbers==1.0.2
scipy==1.15.2
setuptools==79.0.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.3
tiktoken==0.9.0
# torch==2.7.0
tqdm==4.67.1
trio==0.30.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
word2number==1.1
xgboost==3.0.0
xpinyin==0.7.6
yarg==0.1.10
